update: detail.jsp
1. 내가 쓴 글 맨 위에 보이기
2. 만약 회원이 아니거나 리플을 썼을 경우, 입력 창 숨기기
3.  like 기능, 로그인 유무 판별 기능 추가

+++ src/main/webapp/WEB-INF/views/lecture/detail.jsp
@@ -39,6 +39,12 @@
 		console.log("addLike() 호출");
 		const lectureId = $("#lectureId").val();
 		const memberId = $("#memberId").val();
+		
+		if(memberId ===''){
+			alert("로그인을 하셔야 이용하실 수 있습니다.");
+			return;
+		}
+		
 		$.ajax({
 			type : "POST",
 			url : `like/${lectureId}/${memberId}`,
@@ -58,6 +64,11 @@
 		console.log("removeLike() 호출");
 		const lectureId = $("#lectureId").val();
 		const memberId = $("#memberId").val();
+		
+		if(memberId ===''){
+			alert("로그인을 하셔야 이용하실 수 있습니다.");
+			return;
+		}
 		$.ajax({
 			type : "DELETE",
 			url : `like/${lectureId}/${memberId}`,
